fix windows compilation warning

+++ src/base/system.c
@@ -673,7 +673,7 @@ struct THREAD_RUN
 #if defined(CONF_FAMILY_UNIX)
 static void *thread_run(void *user)
 #elif defined(CONF_FAMILY_WINDOWS)
-static unsigned int __stdcall thread_run(void *user)
+static unsigned long __stdcall thread_run(void *user)
 #else
 #error not implemented
 #endif
